⚡️ Static ISR for richlist
